updates to sign up

+++ auction_central/DbWrap.cs
@@ -415,7 +415,7 @@ namespace auction_central
                 */
 
         // create account 
-        public void InsertAdmin(string firstname, string lastname, string email, string password, int phonenumber, Person.UserTypeEnum type)
+        public void InsertAdmin(string firstname, string lastname, string email, string password, Int64 phonenumber, Person.UserTypeEnum type)
         {
             MySqlConnection connection;
             string connectionString = @"Database=auction_central;Data Source=us-cdbr-azure-west-b.cleardb.com;User Id=b1a4a9b19daca1;Password=d28c0eba";
